+++ example/lib/app/routes/app_routes.dart
@@ -1,5 +1,5 @@
 part of 'app_pages.dart';
-// DO NOT EDIT. This is code generated via package:get_cli/get_cli.dart
+// DO NOT EDIT. This is code generated via package:sm_get_cli/sm_get_cli.dart
 
 abstract class Routes {
   Routes._();