+++ lib/common/utils/pubspec/pubspec_lock.dart
@@ -16,7 +16,7 @@ class PubspecLock {
       var pathToPubLock = join(dirname(scriptFile), '../pubspec.lock');
       final file = File(pathToPubLock);
       var text = loadYaml(await file.readAsString());
-      if (text['packages']['get_cli'] == null) {
+      if (text['packages']['sm_get_cli'] == null) {
         if (isDevVersion()) {
           if (!disableLog) {
             LogService.info('Development version');
@@ -24,7 +24,7 @@ class PubspecLock {
         }
         return null;
       }
-      var version = text['packages']['get_cli']['version'].toString();
+      var version = text['packages']['sm_get_cli']['version'].toString();
       return version;
     } on Exception catch (_) {
       if (!disableLog) {