+++ lib/extensions.dart
@@ -1,6 +1,6 @@
 library;
 
-export 'package:get_cli/core/internationalization.dart' show TranslationExt;
-export 'package:get_cli/extensions/dart_code.dart';
-export 'package:get_cli/extensions/list.dart';
-export 'package:get_cli/extensions/string.dart';
+export 'package:sm_get_cli/core/internationalization.dart' show TranslationExt;
+export 'package:sm_get_cli/extensions/dart_code.dart';
+export 'package:sm_get_cli/extensions/list.dart';
+export 'package:sm_get_cli/extensions/string.dart';