+++ lib/functions/formatter_dart_file/frommatter_dart_file.dart
@@ -1,5 +1,5 @@
 import 'package:dart_style/dart_style.dart';
-import 'package:get_cli/common/utils/pubspec/pubspec_utils.dart';
+import 'package:sm_get_cli/common/utils/pubspec/pubspec_utils.dart';
 
 /// Format a dart file
 String formatterDartFile(String content) {