+++ lib/functions/version/version_update.dart
@@ -14,13 +14,15 @@ import 'print_get_cli.dart';
 void checkForUpdate() async {
   if (!CliConfig.updateIsCheckingToday()) {
     if (!isDevVersion()) {
-      await PubDevApi.getLatestVersionFromPackage('get_cli')
+      await PubDevApi.getLatestVersionFromPackage('sm_get_cli')
           .then((versionInPubDev) async {
+        if (versionInPubDev == null) return;
+
         await PubspecLock.getVersionCli(disableLog: true)
             .then((versionInstalled) async {
           if (versionInstalled == null) exit(2);
 
-          final v1 = Version.parse(versionInPubDev!);
+          final v1 = Version.parse(versionInPubDev);
           final v2 = Version.parse(versionInstalled);
           final needsUpdate = v1.compareTo(v2);
           // needs update.