+++ lib/get_cli.dart
@@ -1,3 +1,1 @@
-export 'package:get_cli/commands/impl/commads_export.dart';
-export 'package:get_cli/core/generator.dart';
-export 'package:get_cli/extensions.dart';
+export 'sm_get_cli.dart';