+++ lib/samples/impl/generate_locales.dart
@@ -19,7 +19,7 @@ class GenerateLocalesSample extends Sample {
 
   @override
   String get content => '''
-// DO NOT EDIT. This is code generated via package:get_cli/get_cli.dart
+// DO NOT EDIT. This is code generated via package:sm_get_cli/sm_get_cli.dart
 
 // ignore_for_file: lines_longer_than_80_chars, constant_identifier_names
 // ignore: avoid_classes_with_only_static_members