+++ lib/sm_get_cli.dart
@@ -1,0 +1,3 @@
+export 'package:sm_get_cli/commands/impl/commads_export.dart';
+export 'package:sm_get_cli/core/generator.dart';
+export 'package:sm_get_cli/extensions.dart';