+++ test/fixtures/exception_handler_main.dart
@@ -1,4 +1,4 @@
-import 'package:get_cli/exception_handler/exception_handler.dart';
+import 'package:sm_get_cli/exception_handler/exception_handler.dart';
 
 void main() {
   ExceptionHandler().handle(Exception('fixture failure'));